Added built-in imus to the sequence.
Running it again, its obvious that the onboard imu's are either damaged
(the accelerometer doesn't response to stimuli) or incredibly noisy that
no error correction will fix it without losing actual data. Regardless,
they will be left in as to test with a newer board that may not be as
damaged.

+++ Software/BikeAddon/Core/Inc/xyz_axis.h
@@ -3,9 +3,9 @@
 
 typedef struct
 {
-	int32_t x;
-	int32_t y;
-	int32_t z;
+	double x;
+	double y;
+	double z;
 } xyz_axis_t;
 
 #define IMPLEMENT_READ_XYZ_VALUES(__ID) void read_xyz_values_##__ID(xyz_axis_t* inout_xyz_axis)
